+++ src/main/git/index.ts
@@ -31,6 +31,7 @@ export {
   revertCommit,
   stash,
   stashPop,
+  cloneRepository,
   showExternalDiff,
   openRepoTerminal
 } from './ops'